git-svn-id: https://svn.apache.org/repos/asf/incubator/jena/Import/Jena-SVN/TDB/trunk@1120643 13f79535-47bb-0310-9956-ffa450edef68

+++ src-lib/lib/Bytes.java
@@ -151,7 +151,7 @@ public class Bytes
     /** Make an int order of args -- high to low */
     static private int assembleInt(byte b3, byte b2, byte b1, byte b0)
     {
-        return (int)( ((b3 & 0xFF) << 24) |
+        return ( ((b3 & 0xFF) << 24) |
                       ((b2 & 0xFF) << 16) |
                       ((b1 & 0xFF) <<  8) |
                       ((b0 & 0xFF) <<  0)
